WarriorType enum added. Defender class added.
