+++ TheWarriors/src/main/java/com/warriors/model/Defender.java
@@ -1,0 +1,4 @@
+package com.warriors.model;
+
+public class Defender extends Warrior {
+}
