+++ TheWarriors/src/test/java/com/warriors/ArmyBattleTest.java
@@ -2,8 +2,8 @@ package com.warriors;
 
 import com.warriors.model.Army;
 import com.warriors.model.Knight;
-import com.warriors.model.Unit;
 import com.warriors.model.Warrior;
+import com.warriors.model.WarriorType;
 import com.warriors.service.Battle;
 import org.junit.jupiter.api.Test;
 
@@ -19,8 +19,8 @@ class ArmyBattleTest {
         var knightsArmy = new Army();
         var anotherArmy = new Army();
 
-        attackingArmy.addUnits(Unit.Type.WARRIOR, 2).addUnits(Unit.Type.WARRIOR, 1);
-        defendingArmy.addUnits(Unit.Type.WARRIOR, 3);
+        attackingArmy.addUnits(WarriorType.WARRIOR, 2).addUnits(WarriorType.WARRIOR, 1);
+        defendingArmy.addUnits(WarriorType.WARRIOR, 3);
         Warrior prototype = new Warrior();
         anotherArmy.addUnits(prototype, 2).addUnits(prototype, 1);
         knightsArmy.addUnits(Knight::new, 2).addUnits(Knight::new, 1);
